+++ server-sdk/pom.xml
@@ -1,8 +1,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.amazonaws.ml.mms</groupId>
-  <artifactId>plugins-sdk</artifactId>
+  <groupId>software.amazon.ai</groupId>
+  <artifactId>mms-plugins-sdk</artifactId>
   <packaging>jar</packaging>
   <version>1.0</version>
   <name>server-sdk</name>
