+++ server-sdk/src/main/java/software/amazon/ai/mms/servingsdk/annotations/helpers/EndpointTypes.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.amazonaws.ml.mms.servingsdk.annotations.helpers;
+package software.amazon.ai.mms.servingsdk.annotations.helpers;
 
 /**
  * Types of ModelServer endpoints
